+++ libs/react/fp-button/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.9](https://github.com/shawn-sandy/fp-kit/compare/@fpkit/fp-btn@0.0.8...@fpkit/fp-btn@0.0.9) (2022-07-15)
+
+
+### Features
+
+* ✨ addons for storybook ([8db8492](https://github.com/shawn-sandy/fp-kit/commit/8db84928d6355b3f0edecc726a64199cae7e97f7))
+
+
+
+
+
+## [0.0.8](https://github.com/shawn-sandy/fp-kit/compare/@fpkit/fp-btn@0.0.7...@fpkit/fp-btn@0.0.8) (2022-07-15)
+
+
+### Features
+
+* ✨ add story for button component ([cce442c](https://github.com/shawn-sandy/fp-kit/commit/cce442c8fed0b7d45721a4ae6704e64f2b8b50d2))
+
+
+
+
+
 ## [0.0.7](https://github.com/shawn-sandy/fp-kit/compare/@fpkit/fp-btn@0.0.6...@fpkit/fp-btn@0.0.7) (2022-07-15)